Parser bug fix!
When trying to parse `(let a = 4)` it would result in an error. The
resulting error were unexpected token, since the old way of parsing
declerations would eat all tokens, when trying to close the parenthesize
the parser would encounter EOF, which would result in a parser error.

This fix also allows wierd stuff like `(let a = 4) * 5`, which is
technically allowed, and makes some sense. Everything in this mini
language is an expression, which leads to everything having a value.
This is why you should be able to do this.

But if you were to do something like `let a = 4 * 5`, a would have the
value 20 and in the example above it would have the value 4 al according
to the design of this language.