Fixed bug: caret position will not change when syntax highlighting
When typing in the calculator the syntax highlighting phase won't swap
the carat posiotion. If you type something like: () and the add
something in the middle like (3) and when the '3' is
selected it will now longer put the caret at the end of the expression,
but instead the caret will be after the '3' which is more logical.