Fix: Unaryprecedence of -/+ is now correct.
Since the precedence is allways increasing for the powers we need to
increase the precedence of the unary operators, since we want them to
bind harder than all the binary operators. The soloution set the
precedence of the unaries to be Integer.MAX_VALUE and it will never be
a problem to increase the powers precedence - to make it bind from the
right with power towers, since you wouldn't realisticlly input that many
powers. But if you do then congratulations, you have broken the
expression evaluator.